third_party/WebKit/LayoutTests/virtual/off-main-thread-fetch/http/tests/misc/resource-timing-sizes-xhr-fetch-worker-expected.txt

This is a testharness.js-based test.
FAIL PerformanceResourceTiming sizes XHR and Fetch test assert_equals: seenResources expected 3 but got 1
Harness: the test ran to completion.
